Debugging feedback time diagnostics and throughput drop issue- part 5 investigating CE marking and packet drops part 5 tankreason, CE marks calculation fixed part 6

+++ modules/congestion_controller/l4s/l4s_network_controller.cc
@@ -1667,8 +1667,9 @@ void webrtc::L4SNetworkController::UpdateAckedBitrateEstimator(const TransportPa
 
     if (using_bootstrap && drop_percent > kBootstrapLargeDropThresholdPercent) {
       double guarded_bps = std::max(
-          effective_acked_rate.bps(),
-          last_acked_bitrate_->bps() * (1.0 - kBootstrapMaxDownStepFraction));
+        static_cast<double>(effective_acked_rate.bps()),
+        static_cast<double>(last_acked_bitrate_->bps()) *
+          (1.0 - kBootstrapMaxDownStepFraction));
       smoothed_acked_rate = webrtc::DataRate::BitsPerSec(
           static_cast<int64_t>(guarded_bps));
       bootstrap_drop_guard_applied = true;
